REVIEW/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.25.0
gradio>=3.35.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.28.0
rembg>=2.0.0
imagehash>=4.3.1
scikit-image>=0.19.0 